Preview cell snippet and move CodeSnippetDisplay to its own file
